chore(sequence): rename Series -> Sequence and update stories/exports
