+++ packages/react-components/react-motion-components-preview/library/src/choreography/Sequence/index.ts
@@ -1,0 +1,2 @@
+export { Sequence, Hold, Scene } from './Sequence';
+export type { SequenceProps, HoldProps } from './sequence-types';
