--- packages/react-components/react-motion-components-preview/library/src/choreography/Series/index.ts
@@ -1,2 +1,0 @@
-export { Series, Hold, Scene } from './Series';
-export type { SeriesProps, HoldProps } from './series-types';
