--- packages/react-components/react-motion-components-preview/library/src/choreography/Series/utils/index.ts
@@ -1,3 +1,0 @@
-export { DEFAULT_HOLD_DURATION } from './constants';
-export { getSeriesChildMapping, getSeriesChildArray } from './getSeriesChildMapping';
-export type { SeriesChildMapping } from './getSeriesChildMapping';
