+++ packages/react-components/react-motion-components-preview/stories/src/choreography/Sequence/index.stories.ts
@@ -1,0 +1,15 @@
+import { Sequence } from '@fluentui/react-motion-components-preview';
+
+export { Default } from './SequenceDefault.stories';
+
+export default {
+  title: 'Motion/Choreography (preview)/Sequence',
+  component: Sequence,
+  parameters: {
+    docs: {
+      description: {
+        component: 'Sequence component for sequential animations',
+      },
+    },
+  },
+};
