--- packages/react-components/react-motion-components-preview/stories/src/choreography/Series/index.stories.ts
@@ -1,15 +1,0 @@
-import { Series } from '@fluentui/react-motion-components-preview';
-
-export { Default } from './SeriesDefault.stories';
-
-export default {
-  title: 'Motion/Choreography (preview)/Series',
-  component: Series,
-  parameters: {
-    docs: {
-      description: {
-        component: 'Series component for sequential animations',
-      },
-    },
-  },
-};
